+++ aries-site/src/examples/templates/forms/index.js
@@ -1,4 +1,4 @@
-export * from './do&dont';
+export * from './do-dont';
 export * from './ChangePasswordExample';
 export * from './CustomizeExample';
 export * from './FilterExample';
